renamed napsack.go to Knapsack.go, changed using struct to using weight, value slice, changed printing all to printing max profit.
